+++ pyproject.toml
@@ -1,0 +1,19 @@
+[tool.poetry]
+name = "ssgymshorts"
+version = "0.1.0"
+description = "Process SS Gym logs"
+authors = ["Pradeep Gowda <pradeep@btbytes.com>"]
+license = "MIT"
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.11"
+beautifulsoup4 = "4.12.2"
+simple-parsing = "0.1.4"
+pandas = "2.1.4"
+flask = "^3.0.0"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
