--- requirements.txt
@@ -1,3 +1,0 @@
-beautifulsoup4==4.12.2
-pandas==2.1.4
-simple-parsing==0.1.4
