+++ templates/output.html
@@ -1,0 +1,8 @@
+{% extends 'base.html' %}
+
+{% block title %}Output{% endblock %}
+
+{% block main %}
+<p>TODO: show output here</p>
+<p>TODO: Show download button</p>
+{% endblock %}
